+++ components/common/Layout/Layout.tsx
@@ -4,7 +4,8 @@ import {
 } from '@components/common'
 import { Grid } from '@material-ui/core'
 import { makeStyles } from '@material-ui/core/styles'
-import React from 'react'
+import React, { FC } from 'react'
+import './Layout.module.css'
 
 const useStyles = makeStyles((theme) => ({
   root: {
@@ -17,7 +18,7 @@ const useStyles = makeStyles((theme) => ({
   },
 }))
 
-const Layout = ({ children }) => {
+const Layout: FC = ({ children }) => {
   const classes = useStyles()
 
   return (
