+++ data.js
@@ -1,0 +1,74 @@
+export const Data = [
+  {
+    id: 1,
+    label: 'Home',
+    pathname: '/',
+    navbar: true,
+    homepage: false,
+    icon: '',
+    job: true,
+  },
+  {
+    id: 2,
+    label: 'About',
+    pathname: '/about',
+    navbar: false,
+    homepage: false,
+    icon: '',
+    job: false,
+  },
+  {
+    id: 3,
+    label: 'Resume',
+    pathname:
+      'https://storage.cloud.google.com/sdfdsfasd/Resume-Brian-Lau.pdf',
+    navbar: true,
+    external: true,
+    homepage: true,
+    job: true,
+  },
+  {
+    id: 4,
+    label: 'Development',
+    pathname: '/dev',
+    navbar: true,
+    homepage: true,
+    icon: 'CodeOutlined',
+  },
+  {
+    id: 5,
+    label: 'PhotomoGrapher',
+    pathname: '/photo',
+    navbar: false,
+    homepage: false,
+  },
+  {
+    id: 6,
+    label: 'BodyBuilder',
+    pathname: '/workout',
+    navbar: false,
+    homepage: false,
+  },
+  {
+    id: 7,
+    label: 'Traveller',
+    pathname: '/travel',
+    navbar: false,
+    homepage: false,
+  },
+  {
+    id: 8,
+    label: 'Github',
+    pathname: 'https://github.com/wtlau',
+    external: true,
+    social: true,
+  },
+  {
+    id: 9,
+    label: 'Contact',
+    pathname: '/contact',
+    navbar: true,
+    homepage: true,
+    job: true,
+  },
+]
